requirements.txt
Flask
pymongo
bcrypt
requests
dnspython
python-dotenv
